// frontend/src/api-services/models/video.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ETGY Backend API
 * API documentation generated from Zod schemas (auto sync with validation).
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { VideoMetrics } from './video-metrics';
 /**
 * 
 *
 * @export
 * @interface Video
 */
export interface Video {

    /**
     * @type {number}
     * @memberof Video
     */
    id: number;

    /**
     * @type {string}
     * @memberof Video
     */
    title: string;

    /**
     * @type {string}
     * @memberof Video
     */
    intro?: string | null;

    /**
     * @type {string}
     * @memberof Video
     */
    url: string;

    /**
     * @type {string}
     * @memberof Video
     */
    coverUrl?: string | null;

    /**
     * @type {number}
     * @memberof Video
     */
    duration?: number | null;

    /**
     * @type {string}
     * @memberof Video
     */
    gradeRange?: string | null;

    /**
     * @type {string}
     * @memberof Video
     */
    subjectTag?: string | null;

    /**
     * @type {string}
     * @memberof Video
     */
    status: VideoStatusEnum;

    /**
     * @type {string}
     * @memberof Video
     */
    rejectReason?: string | null;

    /**
     * @type {number}
     * @memberof Video
     */
    reviewedBy?: number | null;

    /**
     * @type {Date}
     * @memberof Video
     */
    reviewedAt?: Date | null;

    /**
     * @type {number}
     * @memberof Video
     */
    publishedBy?: number | null;

    /**
     * @type {Date}
     * @memberof Video
     */
    publishedAt?: Date | null;

    /**
     * @type {number}
     * @memberof Video
     */
    offlineBy?: number | null;

    /**
     * @type {Date}
     * @memberof Video
     */
    offlineAt?: Date | null;

    /**
     * @type {string}
     * @memberof Video
     */
    offlineReason?: string | null;

    /**
     * @type {number}
     * @memberof Video
     */
    uploaderId: number;

    /**
     * @type {number}
     * @memberof Video
     */
    collegeId: number;

    /**
     * @type {Date}
     * @memberof Video
     */
    createdAt: Date;

    /**
     * @type {Date}
     * @memberof Video
     */
    updatedAt: Date;

    /**
     * @type {VideoMetrics}
     * @memberof Video
     */
    metrics?: VideoMetrics;

    /**
     * @type {any}
     * @memberof Video
     */
    uploader?: any | null;

    /**
     * @type {any}
     * @memberof Video
     */
    college?: any | null;
}

/**
 * @export
 * @enum {string}
 */
export enum VideoStatusEnum {
    DRAFT = 'DRAFT',
    REVIEW = 'REVIEW',
    APPROVED = 'APPROVED',
    PUBLISHED = 'PUBLISHED',
    REJECTED = 'REJECTED',
    OFFLINE = 'OFFLINE'
}