// frontend/src/api-services/models/live-room.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ETGY Backend API
 * API documentation generated from Zod schemas (auto sync with validation).
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface LiveRoom
 */
export interface LiveRoom {

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    id: number;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    title: string;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    intro?: string | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    planStartTime: Date;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    planEndTime: Date;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    actualStart?: Date | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    actualEnd?: Date | null;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    status: LiveRoomStatusEnum;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    rejectReason?: string | null;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    reviewedBy?: number | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    reviewedAt?: Date | null;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    publishedBy?: number | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    publishedAt?: Date | null;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    offlineBy?: number | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    offlineAt?: Date | null;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    offlineReason?: string | null;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    pushUrl?: string | null;

    /**
     * @type {string}
     * @memberof LiveRoom
     */
    pullUrl?: string | null;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    replayVideoId?: number | null;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    anchorId: number;

    /**
     * @type {number}
     * @memberof LiveRoom
     */
    collegeId: number;

    /**
     * @type {any}
     * @memberof LiveRoom
     */
    anchor?: any | null;

    /**
     * @type {any}
     * @memberof LiveRoom
     */
    college?: any | null;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    createdAt?: Date;

    /**
     * @type {Date}
     * @memberof LiveRoom
     */
    updatedAt?: Date;
}

/**
 * @export
 * @enum {string}
 */
export enum LiveRoomStatusEnum {
    DRAFT = 'DRAFT',
    REVIEW = 'REVIEW',
    PASSED = 'PASSED',
    PUBLISHED = 'PUBLISHED',
    REJECTED = 'REJECTED',
    LIVING = 'LIVING',
    FINISHED = 'FINISHED',
    OFFLINE = 'OFFLINE'
}